#14 서버현황 API 기능 추가

+++ Ami/src/main/java/com/cnu/ami/device/server/dao/entity/ServerEntity.java
@@ -37,6 +37,6 @@ public class ServerEntity {
 	private long WDATE;
 
 	@Column(name = "RSEQ")
-	private long RSEQ;
+	private int RSEQ;
 
 }
